Inline DoubleClickTracker into the input module
The tracker is used by exactly one caller and is conceptually part of
input state. Folding it in removes a one-struct module file.
